Klt Code Re arranging.
